requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mysql-connector==2.1.6
Werkzeug==0.14.1
